contrib: Remove unnecessary Windows restarts
During minion init, a script is created which will run at startup
and Windows gets rebooted.
Since ovnkube runs as a service, this is not required anymore.

This commit removes the Windows reboot and starts directly the
ovn-kubernetes-node service which will run minion init. The restart
after docker installation is also removed since the path is
already updated.

Signed-off-by: Alin-Gheorghe Balutoiu <alinbalutoiu@gmail.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>
Co-authored-by: Alin Gabriel Serdean <aserdean@ovn.org>